==> analyses/05_outputs.Rmd <==
---
title: "Excess mortality"
subtitle: "Final outputs for paper and appendix"
author: "Radek Panczak"
date: "`r Sys.Date()`"
output:
  rmdformats::robobook:
  code_folding: show
self_contained: true
highlight: pygments
editor_options: 
  chunk_output_type: console
knit: (function(inputFile, encoding) {
  rmarkdown::render(inputFile, encoding = encoding, output_dir = "../docs") })
---

<style type="text/css">
div.page-inner {
max-width: 10000px !important;
}
</style>

<!-- ----------------------------------------------------- -->

```{r setup, include = FALSE}
set.seed(12345)
options(scipen = 999)
options(max.print = "75")

library(pacman)
p_load(tidyverse, magrittr, scales, patchwork, kableExtra, grid)
```

```{r knit-setup, include = FALSE}
knitr::opts_knit$set(root.dir = rprojroot::find_rstudio_root_file())

knitr::opts_chunk$set(cache = FALSE,
                      prompt = FALSE,
                      tidy = FALSE,
                      comment = NA,
                      message = FALSE,
                      warning = FALSE)
knitr::opts_knit$set(width = 125)

mu <- Hmisc::markupSpecs$html
```

<!-- ----------------------------------------------------- -->

# Result files 

```{r include=FALSE}
pandemic_year <- c("1890", "1918", "1957", "2020", "2021")

merge_mort <- function(data){
  
  data %>% 
    mutate(Pandemic = case_when(
      (Year >= 1889 & Year <= 1892) ~ 1890,
      (Year >= 1917 & Year <= 1920) ~ 1918,
      (Year >= 1956 & Year <= 1959) ~ 1957,
      (Year >= 2019 & Year <= 2021) ~ 2020
    )) %>% 
    relocate(Country) %>% 
    mutate(Country = fct_relevel(factor(Country), 
                                 "Switzerland", "Sweden", "Spain"))
}

# custom panel borders
source("R/theme_border.R")
```

## Primary results from age adjusted models

```{r}
results_month <- merge_mort(bind_rows(
  read_rds("data/outputs_2021-11-23/Sweden_results_month.Rds"),
  read_rds("data/outputs_2021-11-24/Spain_results_month.Rds"),
  read_rds("data/outputs_2021-11-23/Switzerland_results_month.Rds")
)) %>% 
  filter(Model == "Age Serfling (Stan, NB)") %>% 
  mutate(Date = anytime::anydate(paste0(Year, "-", Month))) %>% 
  arrange(Country, Year, Month) %>% 
  relocate(c(n_eff, Rhat), .after = last_col()) %>% 
  mutate(across(pred_total_deaths:excess_month_upper, round)) %>% 
  relocate(Model, Pandemic, Country, Year, Month, Date)

results_year <- merge_mort(bind_rows(
  read_rds("data/outputs_2021-11-23/Sweden_results_year.Rds"),
  read_rds("data/outputs_2021-11-24/Spain_results_year.Rds"),
  read_rds("data/outputs_2021-11-23/Switzerland_results_year.Rds") 
)) %>% 
  filter(Model == "Age Serfling (Stan, NB)") %>% 
  arrange(Country, Year) %>% 
  mutate(across(yearly_pred_total_deaths:yearly_excess_total_deaths_upper, round)) %>% 
  relocate(Model, Pandemic)

results_age <- merge_mort(bind_rows(
  read_rds("data/outputs_2021-11-23/Sweden_results_age.Rds"),
  read_rds("data/outputs_2021-11-24/Spain_results_age.Rds"),
  read_rds("data/outputs_2021-11-23/Switzerland_results_age.Rds") 
)) %>% 
  filter(Model == "Age Serfling (Stan, NB)") %>% 
  arrange(Country, Year) %>% 
  mutate(across(pred_grouped_deaths:excess_grouped_lifelost70_upper, round)) %>% 
  relocate(Model, Pandemic, Country, Year, Age_cat, Deaths)
```

```{r include=FALSE}
write_rds(results_month, "data/results_month.Rds")
write_rds(results_year, "data/results_year.Rds")
write_rds(results_age, "data/results_age.Rds")
```

## Sensitivity analysis using Bayesian NB model, unadjusted models

```{r}
results_year_global <- merge_mort(bind_rows(
  read_rds("data/outputs_2021-11-23/Sweden_results_year.Rds"),
  read_rds("data/outputs_2021-11-24/Spain_results_year.Rds"),
  read_rds("data/outputs_2021-11-23/Switzerland_results_year.Rds") 
)) %>% 
  filter(Model == "Global Serfling (Stan, NB)") %>% 
  arrange(Country, Year) %>% 
  mutate(across(yearly_pred_total_deaths:yearly_excess_total_deaths_upper, round)) %>% 
  relocate(Model, Pandemic)
```

## Sensitivity analysis using frequentist Poisson model, unadjusted for age

```{r}
results_month_boot <- read_rds("data/outputs_2021-11-18/All_results_month.Rds") %>% 
  mutate(Date = anytime::anydate(paste0(Year, "-", Month))) %>% 
  arrange(Country, Year, Month) %>% 
  mutate(across(pred:excess_month_lower, round)) %>% 
  mutate(Model = "Global Serfling (bootstrap)") %>% 
  relocate(Model, Country, Year, Month, Date)
```

## Sensitivity analysis using alternative longer (7 days) time window 

```{r}
results_year_last_7_notrim <- merge_mort(bind_rows(
  read_rds("data/outputs_2021-12-01/Sweden_results_year_last_7_notrim.Rds"),
  read_rds("data/outputs_2021-12-01/Spain_results_year_last_7_notrim.Rds"),
  read_rds("data/outputs_2021-12-01/Switzerland_results_year_last_7_notrim.Rds")
)) %>% 
  filter(Model == "Age Serfling (Stan, NB, last 7 no trim)") %>% 
  arrange(Country, Year) %>% 
  mutate(across(yearly_pred_total_deaths:yearly_excess_total_deaths_upper, round)) %>% 
  relocate(Model, Pandemic)
```

## Sensitivity analysis using alternative time window & min/max selection

```{r}
results_year_last_7_trim <- merge_mort(bind_rows(
  read_rds("data/outputs_2021-11-23/Sweden_results_year_last_7_trim.Rds"),
  read_rds("data/outputs_2021-11-24/Spain_results_year_last_7_trim.Rds"),
  read_rds("data/outputs_2021-11-23/Switzerland_results_year_last_7_trim.Rds")
)) %>% 
  filter(Model == "Age Serfling (Stan, NB, last 7)") %>% 
  arrange(Country, Year) %>% 
  mutate(across(yearly_pred_total_deaths:yearly_excess_total_deaths_upper, round)) %>% 
  relocate(Model, Pandemic)
```

## Sensitivity analysis for including effects of pandemics

```{r}
results_month_pand <- merge_mort(bind_rows(
  read_rds("data/outputs_2021-11-23/Sweden_results_month_pand.Rds"),
  read_rds("data/outputs_2021-11-24/Spain_results_month_pand.Rds"),
  read_rds("data/outputs_2021-11-23/Switzerland_results_month_pand.Rds")
)) %>% 
  filter(Model == "Age Serfling (Stan, NB, pandemic)") %>%
  mutate(Date = anytime::anydate(paste0(Year, "-", Month))) %>% 
  relocate(c(n_eff, Rhat), .after = last_col()) %>% 
  arrange(Country, Year, Month) %>% 
  mutate(across(pred_total_deaths:excess_month_upper, round)) %>% 
  relocate(Model, Pandemic, Country, Year, Month, Date)
```

## Population

```{r}
pops <- read_rds("data/deaths_monthly.Rds") %>% 
  filter(Month == 1) %>% 
  select(Year, Country, Population_obs) %>% 
  rename(Population = Population_obs)
```

## COVID-19 cases

```{r}
covid_cases_deaths_jh <- read_rds("data/covid_cases_deaths_jh.Rds")
```

## Yearly deaths by age

```{r}
deaths_yearly_age_sex <- read_rds("data/deaths_yearly_age_sex.Rds")
```

<!-- ----------------------------------------------------- -->

# Paper outputs


## Table 1

Characteristics of the 4 Deadliest Pandemics of the Past 140 Years in Switzerland, Sweden, and Spain, by Calendar Year.  

```{r echo=FALSE}
table_1 <- 
  
  left_join(
    
    # population
    pops %>% 
      filter(Year %in% pandemic_year) %>% 
      select(Year, Country, Population), 
    
    # observed deaths, aggregated from months
    results_month %>% 
      filter(Year %in% pandemic_year) %>% 
      select(Year, Country, Deaths) %>% 
      group_by(Country, Year) %>% 
      summarise(Observed = sum(Deaths))
    
  ) %>% 
  left_join(
    
    results_year %>% 
      filter(Year %in% pandemic_year) %>% 
      mutate(Expected = paste0(number(plyr::round_any(yearly_pred_total_deaths_lower,
                                                      100),
                                      big.mark = " "), 
                               " to ",
                               number(plyr::round_any(yearly_pred_total_deaths_upper,
                                                      100),
                                      big.mark = " "))) %>% 
      select(Year, Country, Expected)
  ) %>% 
  left_join(
    
    covid_cases_deaths_jh) %>% 
  mutate(across(Population:Observed, number, big.mark = " ")) %>% 
  mutate(Deaths = number(Deaths, big.mark = " ")) %>% 
  mutate(across(Observed:Deaths, 
                ~replace(., is.na(.), "--"))) %>% 
  arrange(Year, desc(Country))
```

```{r include=FALSE}
openxlsx::write.xlsx(table_1, "paper/Table_1.xlsx", 
                     rowNames = FALSE, overwrite = TRUE)
```

```{r echo=FALSE}
table_1 %>% 
  select(-Year) %>%
  kbl(align = "lrrrrr",
      col.names = c("", "", 
                    "Count", "Expected", 
                    "Count$^†$")) %>% 
  kable_styling(bootstrap_options = c("striped", "hover", "condensed"),
                full_width = FALSE) %>% 
  add_header_above(c("Calendar Year and Country", "Population, n", 
                     "Deaths From All Causes, n" = 2, 
                     "COVID-19 Deaths, n$^*$" = 1)) %>% 
  pack_rows("1890", 1, 2) %>%
  pack_rows("1918", 3, 5) %>% 
  pack_rows("1957", 6, 8) %>% 
  pack_rows("2020", 9, 11) %>% 
  pack_rows("2021$^‡$", 12, 14) %>% 
  column_spec(1, italic = TRUE) %>%
  row_spec(1, italic = FALSE) %>% 
  footnote(symbol = c("Data are from the Johns Hopkins Coronavirus Resource Center",
                      "Model-predicted number of deaths due to all causes from the 5 preceding years (95% credible interval, rounded to the next 100).",
                      "Only the first 6 months of the year were included in the analyses. No age-specific data were available from Sweden for this period.")) 
```


## Table 2

Model-Based Estimates of Absolute Excess Deaths, Absolute Excess Deaths Scaled by Population, Relative Excess Deaths (Compared With Expected Number of Deaths), and Number of Life-Years Lost Until Age 70 Years.    

```{r echo=FALSE}
table_2 <- 
  
  left_join(
    
    pops %>% 
      filter(Year %in% pandemic_year) %>% 
      select(Year, Country, Population), 
    
    results_year %>%
      select(-starts_with("yearly_pred"), -Model, -Pandemic) %>%
      filter(Year %in% pandemic_year) %>%
      rename(Excess = yearly_excess_total_deaths,
             Excess_lci = yearly_excess_total_deaths_lower, 
             Excess_uci = yearly_excess_total_deaths_upper)
    
  ) %>% 
  
  mutate(
    Percent_excess = 
      number(yearly_rel_excess_total_deaths*100, accuracy = 0.1),
    Percent_lci = 
      number(yearly_rel_excess_total_deaths_lower*100, accuracy = 0.1),
    Percent_uci = 
      number(yearly_rel_excess_total_deaths_upper*100, accuracy = 0.1)
  ) %>% 
  
  mutate(
    Rate_excess = 
      number((Excess / Population) * 100000, 
             accuracy = 5L, big.mark = " "),
    Rate_lci = 
      number((Excess_lci / Population) * 100000, 
             accuracy = 5L, big.mark = " "),
    Rate_uci = 
      number((Excess_uci / Population) * 100000, 
             accuracy = 5L, big.mark = " ")
  ) %>% 
  relocate(Year, Country, 
           starts_with("Excess"), 
           starts_with("Rate"), 
           starts_with("Percent")) %>% 
  
  mutate(Excess = number(Excess, big.mark = " "),
         excess_cri = str_c("(", 
                            number(Excess_lci, big.mark = " "), 
                            " to ", 
                            number(Excess_uci, big.mark = " "), 
                            ")")) %>% 
  relocate(excess_cri, .after = Excess) %>% 
  select(-Excess_lci, -Excess_uci, -starts_with("yearly")) %>% 
  
  mutate(percent_cri = str_c("(", 
                             Percent_lci, 
                             " to ", 
                             Percent_uci, 
                             ")")) %>% 
  relocate(percent_cri, .after = Percent_excess) %>% 
  select(-Percent_lci, -Percent_uci) %>% 
  
  mutate(rate_cri = str_c("(", Rate_lci, " to ", Rate_uci, ")")) %>%   
  relocate(rate_cri, .after = Rate_excess) %>% 
  select(-Rate_lci, -Rate_uci, -Population) %>% 
  arrange(Year, desc(Country))
```

```{r include=FALSE}
openxlsx::write.xlsx(table_2, "paper/Table_2.xlsx", 
                     rowNames = FALSE, overwrite = TRUE)
```

```{r echo=FALSE}
table_2 %>% 
  filter(!(Country == "Sweden" & Year == 2021)) %>% 
  mutate(percent_cri = if_else(Year==2020,paste0(percent_cri,"$^§$"),percent_cri)) %>% 
  select(-Year) %>% 
  kbl(align = "lrrrrrr",
      col.names = c("", 
                    "n", "95% CrI",
                    "n", "95% CrI", 
                    "%", "95% CrI")) %>% 
  kable_styling(bootstrap_options = c("striped", "hover", "condensed"),
                full_width = FALSE) %>% 
  add_header_above(c("Calendar Year and Country" = 1, 
                     "Absolute Excess Deaths per\n100 000 Persons " = 2,
                     "Absolute excess deaths\nper population$^*$" = 2,
                     "Relative Excess Deaths$^†$" = 2
  )) %>% 
  pack_rows("1890", 1, 2) %>%
  pack_rows("1918", 3, 5) %>% 
  pack_rows("1957", 6, 8) %>% 
  pack_rows("2020", 9, 11) %>% 
  pack_rows("2021", 12, 13) %>% 
  column_spec(1, italic = TRUE) %>%
  row_spec(1, italic = FALSE) %>% 
  footnote(symbol=c("Absolute number of excess deaths scaled by country population.",
                    "Absolute number of excess deaths scaled by the number of expected deaths.",
                    "Sum of life years lost until the age of 70 in thousands.",
                    "Based on 8000 posterior samples, the probability that the relative number of deaths in 2020 was larger than in 1957 is 100% for Switzerland, 94.3% for Sweden, and 99.9% for Spain."))
```


## Figure 1

Yearly excess per population.  

**Note:** *the scales of Y axes vary by country!*

```{r echo=FALSE}
vline_pandemic <- tibble(Year=c(1890, 1918, 1957, 1968, 1977, 2009, 2020, 2021)) 

results_year %>% 
  filter(Year >= 1880) %>% 
  filter(Year <= 2021) %>% 
  left_join(pops) %>% 
  mutate(Country = fct_relevel(factor(Country), "Switzerland", "Sweden", "Spain")) %>% 
  mutate(Difference = ifelse(yearly_excess_total_deaths > 0, "More than expected", "Fewer than expected"),
         abs_excess_pop = yearly_excess_total_deaths / Population*100000) %>% 
  ggplot(aes(x = Year)) +
  geom_vline(data = vline_pandemic, 
             aes(xintercept = Year), 
             size = 2, colour = "grey90", alpha = 0.7) +  
  geom_hline(yintercept = 0, colour = "grey80") +
  geom_col(aes(y = abs_excess_pop, fill = Difference)) +
  scale_x_continuous(
    breaks = c(1890, 1918, 1957, 1968, 1977, 2009, 2020),
    limits = c(1880-1, 2022+1),
    expand = c(0, 0)) +
  scale_fill_manual(values = c("#67A9CF", "#EF8A62")) +  
  facet_wrap(facets = vars(Country), 
             nrow = 3,
             scales = "free_y",
             strip.position = "top") + 
  theme_bw() +
  theme(
    strip.background = element_blank(),
    strip.text.x = element_text(margin = margin(0, 0, 0, 0)),
    panel.background = element_blank(), 
    panel.border = theme_border(type = c("bottom", "left")),
    axis.line = element_line(colour = "grey40"),
    axis.ticks = element_line(),
    panel.grid.major = element_blank(), 
    panel.grid.minor = element_blank(),
    legend.position = "bottom",
    legend.background = element_blank(),
    legend.title = element_blank(),
    legend.margin = margin(0, 0, 0, 0),
    legend.box.margin = margin(-20, -10, -10, -10),
    legend.key.size = unit(0.25, "cm"),
    plot.margin = margin(3, 0, 10, 0)) +
  xlab("") + ylab("Absolute Excess Deaths per 100 000 Persons, n") +
  guides(fill = guide_legend(reverse = TRUE, nrow = 1))

rm(vline_pandemic)
```

```{r include=FALSE}
ggsave("paper/Figure_1.png", 
       dpi = 300, width = 190, height = 140, units = "mm")
```

## Figure 2 

Zooming to pandemics to explore absolute excess with CrI.  

```{r include=FALSE}
focus_years <- c(1890, 1918, 1957, 2020)
```

```{r figure2, echo=FALSE}
shown_years = NULL

for(i in focus_years) {
  shown_years = bind_rows(shown_years,
                          tibble(Year = (i-2):(i+2),
                                 focus_year = i,
                                 shown_years=paste0(i-2, 
                                                    "-", 
                                                    min(2020, i+2))))
}

shown_years %<>% filter(Year <= 2021) %>% 
  mutate(shown_years = ifelse(shown_years == "2018-2020", 
                              "2018-2021", shown_years))

plot_data = shown_years %>%
  left_join(results_month) %>% 
  mutate(Country = fct_relevel(factor(Country), "Switzerland", "Sweden", "Spain")) %>% 
  filter(!(Country == "Sweden" & Year == 2021)) %>%
  mutate(focus_year1 = as.Date(paste0(focus_year, "-01-01")),
         focus_year2 = as.Date(paste0(focus_year, "-12-31")),
         focus_year2 = if_else(focus_year==2020, as.Date(paste0(focus_year+1, "-06-30")), focus_year2)) %>% 
  mutate(customlim=case_when(Country=="Switzerland" ~ 330,
                             Country=="Sweden" ~ 330,
                             Country=="Spain" ~ 900))


gg = ggplot(plot_data) +
  geom_point(aes(x=Date,y=customlim),size=.1,colour="transparent") +
  geom_rect(aes(xmin = focus_year1,
                xmax = focus_year2,
                ymin = -Inf, ymax = Inf), 
            fill = "grey90") +
  geom_ribbon(aes(x = Date, 
                  ymin = pred_total_deaths_lower/Population*100000,
                  ymax = pred_total_deaths_upper/Population*100000),
              fill= "aquamarine3", alpha = 0.5) +
  geom_line(aes(x = Date,
                y = pred_total_deaths/Population*100000),
            colour = "aquamarine4") +
  geom_line(aes(x = Date,
                y = Deaths/Population*100000),
            colour = "#EF8A62", size = 0.5) +
  # geom_hline(aes(yintercept=0),colour="black",alpha=.4,linetype=2) +
  geom_ribbon(aes(x = Date,
                  ymin=0,
                  ymax = (Deaths-pred_total_deaths)/Population*100000),
              fill = "cornflowerblue", size = 0, alpha=.4) +
  geom_point(data = shown_years,
             aes(x = as.Date(paste0(Year, "-01-01")),
                 y = 0),
             colour = "transparent") +
  facet_grid(Country ~ shown_years,scales="free", spac="free_x") +
  # scale_y_continuous(labels = label_number(accuracy = 1, suffix = "K", 
  # scale = 1e-3)) +
  scale_x_date(date_breaks = "2 years", date_labels = "%Y") +
  theme_bw()+
  theme(strip.background = element_blank(),
        strip.text.x = element_text(margin = margin(0, 0, 2, 0)),
        strip.text.y = element_text(margin = margin(0, 0, 2, 0)),
        plot.margin = margin(0, 0, -10, 0),
        panel.border = theme_border(type = c("bottom", "left")),
        panel.grid.major = element_blank(), 
        panel.grid.minor = element_blank()) +
  labs(x = "", y = "Monthly Deaths per 100,000 Persons, n")

# gg

g_data_grob = ggplotGrob(gg)

# View(g_data_grob$layout)
# gtable::gtable_show_layout(ggplotGrob(gg)) 

## remove unused facets
idx <- which(g_data_grob$layout$name %in% c("panel-3-1"))
for (i in idx) g_data_grob$grobs[[i]] <- nullGrob()

## move y axis right
# axis-l-1 needs to move 1 column to the right
idx <- which(g_data_grob$layout$name %in% c("axis-l-3"))
g_data_grob$layout[idx, c("l", "r")] <- g_data_grob$layout[idx, c("l", "r")] + 2

## move x axis up
# axis-b-1 needs to move 1 column up
idx <- which(g_data_grob$layout$name %in% c("axis-b-1"))
g_data_grob$layout[idx, c("t")] <- g_data_grob$layout[idx, c("t")] - 2

grid.newpage()
grid.draw(g_data_grob)

gg = ggplotify::as.ggplot(g_data_grob)
```

```{r include=FALSE}
ggsave(plot = gg, "paper/Figure_2.png", 
       dpi = 300, width =  190, height = 140, units = "mm")
```

## Figure 3

Age differences across major pandemics (per population).  

**Note:** *Excludes category `[90,120]` in pandemic years `1890` & `1918` due to small number of deaths in these age groups resulting in very large CrI.*  

```{r include=FALSE}
pandemic_years <- c(1890, 1918, 1957, 2020, 2021)
```

```{r figure3, echo=FALSE}
gap = .3

results_age %>% 
  filter(Year %in% pandemic_years) %>%
  filter(!(Year<1957 & Age_cat=="[90,120]")) %>% 
  mutate(Age_cat2 = as.numeric(Age_cat)) %>%
  mutate(Age_cat2 = case_when(Country == "Spain" ~ Age_cat2-gap,
                              Country == "Sweden" ~ Age_cat2,
                              Country == "Switzerland" ~ Age_cat2+gap)) %>% 
  left_join(deaths_yearly_age_sex) %>%
  ggplot() +
  geom_line(aes(x = Age_cat2,
                y = excess_grouped_deaths/Population_obs*100000,
                colour = factor(Country))) +
  geom_pointrange(aes(x = Age_cat2,
                      y = excess_grouped_deaths/Population_obs*100000,
                      ymin = excess_grouped_deaths_lower/Population_obs*100000,
                      ymax = excess_grouped_deaths_upper/Population_obs*100000,
                      colour = factor(Country)),
                  size = .25) +
  
  geom_hline(yintercept = 0, linetype = 2, colour = "grey90") +
  facet_wrap(~Year, nrow = 2) +
  scale_x_continuous(labels = unique(as.character(results_age$Age_cat)),
                     breaks = 1:10) +
  scale_y_continuous(labels = function(x) number(x, big.mark=",")) +
  theme_bw() +
  scale_colour_manual(values = c("darkgoldenrod1", "dodgerblue", "firebrick2")) +
  coord_cartesian(ylim = c(-600, 5000)) +
  labs(x = "Year",
       y = "Montly deaths") +
  theme(legend.position = c(.85, .2),
        legend.background = element_blank(),
        strip.background = element_blank(),
        strip.text.x = element_text(margin = margin(0, 0, 2, 0)),
        axis.text.x = element_text(angle = 45, hjust = 1),
        panel.border = theme_border(type = c("bottom", "left")),
        panel.grid.major = element_blank(), 
        panel.grid.minor = element_blank(),
        plot.margin = margin(0, 0, 0, 0)) +
  labs(x = "Age group, y",
       y = "Absolute Excess Deaths per\n100,000 Persons, n",
       colour = "Country")
```

```{r include=FALSE}
ggsave("paper/Figure_3.png", 
       dpi = 300, width = 190, height = 140, units = "mm")
```

<!-- ----------------------------------------------------- -->

# Appendix 

## Table S1 

Comparing main estimates (`Age Serfling (Stan, NB)`) of excess from alternative methods: 

- `Global Serfling (bootstrap)`  
- `Global Serfling (Stan, NB)`  
- `Age Serfling with alternative 7 day time window and min/max trim (Stan, NB)`  
- `Age Serfling with alternative 7 day time window and no trim (Stan, NB)`  

```{r}
pandemic_years <- c(1890, 1918, 1957, 2020, 2021)

table_s1 <- 
  
  bind_rows(
    
    # baseline, ie. reported in paper `Age Serfling (Stan, NB)`
    results_year  %>% 
      mutate(Model = "m0") %>% 
      select(Model, Country, Year,
             starts_with("yearly_excess_total_deaths")),
    
    # Global Serfling (bootstrap)
    # simply summed from monthly values to obtain year
    # !without CI for yearly!
    results_month_boot %>% 
      mutate(excess = Deaths - pred) %>% 
      group_by(Model, Country, Year) %>% 
      summarise(yearly_excess_total_deaths = sum(excess)) %>% 
      ungroup() %>% 
      mutate(Model = "m1"),
    
    # Global Serfling (Stan, NB) without age
    results_year_global  %>% 
      mutate(Model = "m2") %>% 
      select(Model, Country, Year,
             starts_with("yearly_excess_total_deaths")),
    
    # Age Serfling with alternative time window & min/max trim (Stan, NB)
    results_year_last_7_notrim  %>% 
      mutate(Model = "m3") %>% 
      select(Model, Country, Year,
             starts_with("yearly_excess_total_deaths")),
    
    # Age Serfling with alternative time window (Stan, NB)
    results_year_last_7_trim  %>% 
      mutate(Model = "m4") %>% 
      select(Model, Country, Year,
             starts_with("yearly_excess_total_deaths"))
    
  ) %>% 
  filter(Year %in% c(pandemic_years, pandemic_years - 1, pandemic_years + 1)) %>% 
  arrange(desc(Country), Year, Model) %>% 
  pivot_wider(names_from = Model, 
              values_from = starts_with("yearly_excess"),
              names_sort = FALSE) %>% 
  relocate(Country, Year,
           yearly_excess_total_deaths_m0, 
           yearly_excess_total_deaths_lower_m0, 
           yearly_excess_total_deaths_upper_m0,
           yearly_excess_total_deaths_m1, 
           yearly_excess_total_deaths_lower_m1, 
           yearly_excess_total_deaths_upper_m1,
           yearly_excess_total_deaths_m2, 
           yearly_excess_total_deaths_lower_m2, 
           yearly_excess_total_deaths_upper_m2,
           yearly_excess_total_deaths_m3, 
           yearly_excess_total_deaths_lower_m3, 
           yearly_excess_total_deaths_upper_m3,
           yearly_excess_total_deaths_m4, 
           yearly_excess_total_deaths_lower_m4, 
           yearly_excess_total_deaths_upper_m4
  ) %>% 
  select(-yearly_excess_total_deaths_lower_m1, -yearly_excess_total_deaths_upper_m1) %>% 
  
  mutate(
    yearly_excess_total_deaths_m0 = number(yearly_excess_total_deaths_m0, 
                                           accuracy = 5L, big.mark = " "), 
    yearly_excess_total_deaths_m1 = number(yearly_excess_total_deaths_m1, 
                                           accuracy = 5L, big.mark = " "),
    yearly_excess_total_deaths_m2 = number(yearly_excess_total_deaths_m2, 
                                           accuracy = 5L, big.mark = " "),
    yearly_excess_total_deaths_m3 = number(yearly_excess_total_deaths_m3, 
                                           accuracy = 5L, big.mark = " "),
    yearly_excess_total_deaths_m4 = number(yearly_excess_total_deaths_m4, 
                                           accuracy = 5L, big.mark = " ")
  ) %>% 
  
  mutate(
    yearly_excess_total_deaths_m0_cri = str_c("(",
                                              number(yearly_excess_total_deaths_lower_m0,
                                                     accuracy = 5L, big.mark = " "), 
                                              " to ",
                                              number(yearly_excess_total_deaths_upper_m0,
                                                     accuracy = 5L, big.mark = " "), 
                                              ")")
  ) %>% 
  relocate(yearly_excess_total_deaths_m0_cri, 
           .after = yearly_excess_total_deaths_m0) %>% 
  select(-yearly_excess_total_deaths_lower_m0,
         -yearly_excess_total_deaths_upper_m0) %>% 
  
  mutate(
    yearly_excess_total_deaths_m2_cri = str_c("(",
                                              number(yearly_excess_total_deaths_lower_m2,
                                                     accuracy = 5L, big.mark = " "), 
                                              " to ",
                                              number(yearly_excess_total_deaths_upper_m2,
                                                     accuracy = 5L, big.mark = " "), 
                                              ")")
  ) %>% 
  relocate(yearly_excess_total_deaths_m2_cri, 
           .after = yearly_excess_total_deaths_m2) %>% 
  select(-yearly_excess_total_deaths_lower_m2,
         -yearly_excess_total_deaths_upper_m2) %>% 
  
  mutate(
    yearly_excess_total_deaths_m3_cri = str_c("(",
                                              number(yearly_excess_total_deaths_lower_m3,
                                                     accuracy = 5L, big.mark = " "), 
                                              " to ",
                                              number(yearly_excess_total_deaths_upper_m3,
                                                     accuracy = 5L, big.mark = " "), 
                                              ")")
  ) %>% 
  relocate(yearly_excess_total_deaths_m3_cri, .after = yearly_excess_total_deaths_m3) %>% 
  select(-yearly_excess_total_deaths_lower_m3, -yearly_excess_total_deaths_upper_m3) %>% 
  
  mutate(
    yearly_excess_total_deaths_m4_cri = str_c("(", 
                                              number(yearly_excess_total_deaths_lower_m4,
                                                     accuracy = 5L, big.mark = " "), 
                                              " to ",
                                              number(yearly_excess_total_deaths_upper_m4,
                                                     accuracy = 5L, big.mark = " "), 
                                              ")")
  ) %>% 
  relocate(yearly_excess_total_deaths_m4_cri, .after = yearly_excess_total_deaths_m4) %>% 
  select(-yearly_excess_total_deaths_lower_m4, -yearly_excess_total_deaths_upper_m4) %>% 
  
  mutate(across(yearly_excess_total_deaths_m0:yearly_excess_total_deaths_m4_cri, 
                ~replace(., is.na(.), "--")))
```

```{r include=FALSE}
openxlsx::write.xlsx(table_s1, "paper/Table_S1.xlsx", 
                     rowNames = FALSE, overwrite = TRUE)
```

```{r echo=FALSE}
table_s1 %>% 
  select(-Country) %>% 
  kbl(align = "lrrrrrrrrr",
      col.names = c("Country, Year", 
                    "Excess", "95% CrI", 
                    "Excess", 
                    "Excess", "95% CrI", 
                    "Excess", "95% CrI", 
                    "Excess", "95% CrI")) %>% 
  add_header_above(c(" " = 1, 
                     "Baseline" = 2, 
                     "Sensitivity 1$^*$" = 1, 
                     "Sensitivity 2$^†$" = 2, 
                     "Sensitivity 3$^‡$" = 2,  
                     "Sensitivity 4$^§$" = 2)) %>%   
  kable_styling(bootstrap_options = c("striped", "hover", "condensed"),
                full_width = FALSE) %>% 
  column_spec(1, italic = TRUE) %>%
  row_spec(1, italic = FALSE) %>% 
  pack_rows("Switzerland", 1, 12) %>%
  pack_rows("Sweden", 13, 24) %>% 
  pack_rows("Spain", 25, 33) %>% 
  footnote(symbol=c("Serfling, Poisson, unadjusted for age",
                    "Serfling, NB, unadjusted for age",
                    "Serfling, NB, age adjusted, alternative time window, no trim",
                    "Serfling, NB, age adjusted, alternative time window, trim"))
```

```{r eval=FALSE, include=FALSE}
table_s1 %>% 
  select(-Country) %>% 
  kbl(align = "lrrrrrrrr",
      col.names = c("Country, Year", 
                    "Excess", "95% CrI", 
                    "Excess", 
                    "Excess", "95% CrI", 
                    "Excess", "95% CrI", 
                    "Excess", "95% CrI")) %>% 
  add_header_above(c(" " = 1, 
                     "Baseline" = 2, 
                     "Sensitivity 1*" = 1, 
                     "Sensitivity 2†" = 2, 
                     "Sensitivity 3‡" = 2,  
                     "Sensitivity 4§" = 2)) %>%   
  kable_styling(bootstrap_options = c("striped", "hover", "condensed"),
                full_width = FALSE) %>% 
  column_spec(1, italic = TRUE) %>%
  row_spec(1, italic = FALSE) %>% 
  pack_rows("Switzerland", 1, 12) %>%
  pack_rows("Sweden", 13, 24) %>% 
  pack_rows("Spain", 25, 33) %>% 
  footnote(symbol=c("Serfling, Poisson, unadjusted for age",
                    "Serfling, NB, unadjusted for age",
                    "Serfling, NB, age adjusted, alternative time window, no trim",
                    "Serfling, NB, age adjusted, alternative time window, trim")) %>% 
  as_image(file = "paper/Table_S1.png")
```

## Figure S1 {.tabset}

Yearly estimates of excess

### Absolute 

```{r echo=FALSE}
vline_pandemic <- tibble(Year=c(1890, 1918, 1957, 1968, 1977, 2009, 2020)) 

results_year %>% 
  filter(Year >= 1880) %>% 
  filter(Year <= 2021) %>% 
  left_join(pops) %>% 
  mutate(Country = fct_relevel(factor(Country), "Switzerland", "Sweden", "Spain")) %>% 
  mutate(Difference = ifelse(yearly_excess_total_deaths > 0, "More than expected", "Fewer than expected"),
         abs_excess_pop = yearly_excess_total_deaths) %>% 
  ggplot(aes(x = Year)) +
  geom_vline(data = vline_pandemic, 
             aes(xintercept = Year), 
             size = 2, colour = "grey90", alpha = 0.7) +  
  geom_hline(yintercept = 0, colour = "grey80") +
  geom_col(aes(y = abs_excess_pop, fill = Difference)) +
  scale_x_continuous(
    breaks = c(1890, 1918, 1957, 1968, 1977, 2009, 2020),
    limits = c(1880-1, 2022+1),
    expand = c(0, 0)) +
  scale_fill_manual(values = c("#67A9CF", "#EF8A62")) +  
  facet_wrap(facets = vars(Country), 
             nrow = 3,
             scales = "free_y",
             strip.position = "top") + 
  theme_bw() +
  theme(
    strip.background = element_blank(),
    strip.text.x = element_text(margin = margin(0, 0, 0, 0)),
    panel.background = element_blank(), 
    panel.border = theme_border(type = c("bottom", "left")),
    axis.line = element_line(colour = "grey40"),
    axis.ticks = element_line(),
    panel.grid.major = element_blank(), 
    panel.grid.minor = element_blank(),
    legend.position = "bottom",
    legend.background = element_blank(),
    legend.title = element_blank(),
    legend.margin = margin(0, 0, 0, 0),
    legend.box.margin = margin(-20, -10, -10, -10),
    legend.key.size = unit(0.25, "cm"),
    plot.margin = margin(3, 0, 10, 0)) +
  xlab("") + ylab("Absolute Excess Deaths, n") +
  guides(fill = guide_legend(reverse = TRUE, nrow = 1))
```

```{r include=FALSE}
ggsave("paper/Figure_S1a.png", dpi = 300, width = 190, height = 140, units = "mm")
```

### Percentage   

```{r echo=FALSE}
results_year %>% 
  filter(Year >= 1880) %>% 
  filter(Year <= 2021) %>% 
  left_join(pops) %>% 
  mutate(Country = fct_relevel(factor(Country), "Switzerland", "Sweden", "Spain")) %>% 
  mutate(Difference = ifelse(yearly_excess_total_deaths > 0, "More than expected", "Fewer than expected"),
         abs_excess_pop = yearly_rel_excess_total_deaths) %>% 
  ggplot(aes(x = Year)) +
  geom_vline(data = vline_pandemic, 
             aes(xintercept = Year), 
             size = 2, colour = "grey90", alpha = 0.7) +  
  geom_hline(yintercept = 0, colour = "grey80") +
  geom_col(aes(y = abs_excess_pop, fill = Difference)) +
  scale_x_continuous(
    breaks = c(1890, 1918, 1957, 1968, 1977, 2009, 2020),
    limits = c(1880-1, 2022+1),
    expand = c(0, 0)) +
  scale_y_continuous(labels = percent) +
  scale_fill_manual(values = c("#67A9CF", "#EF8A62")) +  
  facet_wrap(facets = vars(Country), 
             nrow = 3,
             scales = "free_y",
             strip.position = "top") + 
  theme_bw() +
  theme(
    strip.background = element_blank(),
    strip.text.x = element_text(margin = margin(0, 0, 0, 0)),
    panel.background = element_blank(), 
    panel.border = theme_border(type = c("bottom", "left")),
    axis.line = element_line(colour = "grey40"),
    axis.ticks = element_line(),
    panel.grid.major = element_blank(), 
    panel.grid.minor = element_blank(),
    legend.position = "bottom",
    legend.background = element_blank(),
    legend.title = element_blank(),
    legend.margin = margin(0, 0, 0, 0),
    legend.box.margin = margin(-20, -10, -10, -10),
    legend.key.size = unit(0.25, "cm"),
    plot.margin = margin(3, 0, 10, 0)) +
  xlab("") + ylab("Absolute Excess Deaths, %") +
  guides(fill = guide_legend(reverse = TRUE, nrow = 1))

rm(vline_pandemic)
```

```{r include=FALSE}
ggsave("paper/Figure_S1b.png", 
       dpi = 300, width = 190, height = 140, units = "mm")
```

## Figure S2 {.tabset}

Monthly estimates of excess

### Absolute 

```{r include=FALSE}
vline_pandemic <- tibble(Date=c(as.Date("1890-01-01"),
                                as.Date("1918-01-01"),
                                as.Date("1957-01-01"),
                                as.Date("1968-01-01"),
                                as.Date("1977-01-01"),
                                as.Date("2009-01-01"),
                                as.Date("2020-01-01")))

```

```{r echo=FALSE}
hline_pandemic <- results_month %>% 
  filter(Year >= 2020) %>% 
  group_by(Country) %>% 
  summarise(yintercept = max(excess_month)) %>% 
  ungroup() 

results_month %>% 
  filter(Year >= 1880) %>% 
  mutate(Difference = ifelse(excess_month > 0, "More than expected", "Fewer than expected")) %>% 
  ggplot(aes(x = Date)) +
  geom_vline(data = vline_pandemic, 
             aes(xintercept = Date+365/2), 
             size = 2, colour = "grey90") +
  geom_hline(data = hline_pandemic, 
             aes(yintercept = yintercept), 
             colour = "#984ea3", size = 0.25) +
  geom_col(aes(y = excess_month, fill = Difference, colour = Difference), width = 1) +
  geom_hline(yintercept = 0, linetype = 2, size = 0.25) +
  scale_x_date(date_labels = "%Y",
               breaks = seq(as.Date("1880-01-01"),
                            as.Date("2020-12-31"), "10 years"),
               limits = c(as.Date("1879-12-31"), 
                          as.Date("2021-07-01")), 
               expand = c(0.01, 0)) +
  scale_fill_manual(values = c("#67A9CF", "#EF8A62")) +  
  scale_color_manual(values = c("#67A9CF", "#EF8A62"), guide = "none") +
  facet_grid(Country~., scales = "free_y") + 
  scale_y_continuous(labels = label_number(accuracy = 1, suffix = "K", scale = 1e-3)) +
  facet_wrap(facets = vars(Country), 
             nrow = 3,
             scales = "free_y",
             strip.position = "top") + 
  ylab("Monthly Excess Deaths, n") + 
  theme_bw() +
  theme(
    strip.background = element_blank(),
    strip.text.x = element_text(margin = margin(0, 0, 0, 0)),
    panel.background = element_blank(), 
    panel.border = theme_border(type = c("bottom", "left")),
    axis.line = element_line(colour = "grey40"),
    axis.ticks = element_line(),
    panel.grid.major = element_blank(), 
    panel.grid.minor = element_blank(),
    legend.position = "bottom",
    legend.background = element_blank(),
    legend.title = element_blank(),
    legend.margin = margin(0, 0, 0, 0),
    legend.box.margin = margin(-20, -10, -10, -10),
    legend.key.size = unit(0.25, "cm"),
    plot.margin = margin(3, 0, 10, 0)) +
  xlab("") +
  guides(fill = guide_legend(reverse = TRUE, nrow = 1))
```

```{r include=FALSE}
ggsave("paper/Figure_S2a.png", 
       dpi = 300, width = 190, height = 140, units = "mm")
```

### Relative 

```{r echo=FALSE}
hline_pandemic <- results_month %>% 
  filter(Year >= 2020) %>% 
  group_by(Country) %>% 
  summarise(yintercept = max(excess_month/Population*100000)) %>% 
  ungroup() 

results_month %>% 
  filter(Year >= 1880) %>% 
  mutate(Difference = ifelse(excess_month > 0, "More than expected", "Fewer than expected")) %>% 
  ggplot(aes(x = Date)) +
  geom_vline(data = vline_pandemic, 
             aes(xintercept = Date+365/2), 
             size = 2, colour = "grey90") +
  geom_hline(data = hline_pandemic, 
             aes(yintercept = yintercept), 
             colour = "#984ea3", size = 0.25) +
  geom_col(aes(y = excess_month/Population*100000, 
               fill = Difference, colour = Difference), width = 1) +
  geom_hline(yintercept = 0, linetype = 2, size = 0.25) +
  scale_x_date(date_labels = "%Y",
               breaks = seq(as.Date("1880-01-01"),
                            as.Date("2020-12-31"), "10 years"),
               limits = c(as.Date("1879-12-31"), 
                          as.Date("2021-07-01")), 
               expand = c(0.01, 0)) +
  scale_fill_manual(values = c("#67A9CF", "#EF8A62")) +  
  scale_color_manual(values = c("#67A9CF", "#EF8A62"), guide = "none") +
  facet_wrap(facets = vars(Country), 
             nrow = 3,
             scales = "free_y",
             strip.position = "top") + 
  ylab("Monthly Excess Deaths per 100 000 Persons, n") + 
  theme_bw() +
  theme(
    strip.background = element_blank(),
    strip.text.x = element_text(margin = margin(0, 0, 0, 0)),
    panel.background = element_blank(), 
    panel.border = theme_border(type = c("bottom", "left")),
    axis.line = element_line(colour = "grey40"),
    axis.ticks = element_line(),
    panel.grid.major = element_blank(), 
    panel.grid.minor = element_blank(),
    legend.position = "bottom",
    legend.background = element_blank(),
    legend.title = element_blank(),
    legend.margin = margin(0, 0, 0, 0),
    legend.box.margin = margin(-20, -10, -10, -10),
    legend.key.size = unit(0.25, "cm"),
    plot.margin = margin(3, 0, 10, 0)) +
  xlab("") +
  guides(fill = guide_legend(reverse = TRUE, nrow = 1))

rm(hline_pandemic)
```

```{r include=FALSE}
ggsave("paper/Figure_S2b.png", 
       dpi = 300, width = 190, height = 140, units = "mm")
```

### Percentage 

```{r echo=FALSE}
hline_pandemic <- results_month %>% 
  filter(Year >= 2020) %>% 
  group_by(Country) %>% 
  summarise(yintercept = max(rel_excess_month)) %>% 
  ungroup() 

results_month %>% 
  filter(Year >= 1880) %>% 
  mutate(Difference = ifelse(rel_excess_month > 0, "More than expected", "Fewer than expected")) %>% 
  ggplot(aes(x = Date)) +
  geom_vline(data = vline_pandemic, 
             aes(xintercept = Date+365/2), 
             size = 2, colour = "grey90") +
  geom_hline(data = hline_pandemic, 
             aes(yintercept = yintercept), 
             colour = "#984ea3", size = 0.25) +
  geom_col(aes(y = rel_excess_month, fill = Difference, colour = Difference), width = 1) +
  geom_hline(yintercept = 0, linetype = 2, size = 0.25) +
  scale_x_date(date_labels = "%Y",
               breaks = seq(as.Date("1880-01-01"),
                            as.Date("2020-12-31"), "10 years"),
               limits = c(as.Date("1879-12-31"), 
                          as.Date("2021-07-01")), 
               expand = c(0.01, 0)) +
  scale_fill_manual(values = c("#67A9CF", "#EF8A62")) +  
  scale_color_manual(values = c("#67A9CF", "#EF8A62"), guide = "none") +
  scale_y_continuous(labels = percent) +
  facet_wrap(facets = vars(Country), 
             nrow = 3,
             scales = "free_y",
             strip.position = "top") + 
  ylab("Monthly Excess Deaths, %") + 
  theme_bw() +
  theme(
    strip.background = element_blank(),
    strip.text.x = element_text(margin = margin(0, 0, 0, 0)),
    panel.background = element_blank(), 
    panel.border = theme_border(type = c("bottom", "left")),
    axis.line = element_line(colour = "grey40"),
    axis.ticks = element_line(),
    panel.grid.major = element_blank(), 
    panel.grid.minor = element_blank(),
    legend.position = "bottom",
    legend.background = element_blank(),
    legend.title = element_blank(),
    legend.margin = margin(0, 0, 0, 0),
    legend.box.margin = margin(-20, -10, -10, -10),
    legend.key.size = unit(0.25, "cm"),
    plot.margin = margin(3, 0, 10, 0)) +
  xlab("") +
  guides(fill = guide_legend(reverse = TRUE, nrow = 1))

rm(hline_pandemic)
```

```{r include=FALSE}
ggsave("paper/Figure_S2c.png", 
       dpi = 300, width = 190, height = 140, units = "mm")
```

## Figure S3 {.tabset}

Zooming to pandemics to explore excess

### Absolute

```{r include=FALSE}
plot_years <- c(seq(1890 - 1, 1890 + 2),
                seq(1918 - 1, 1918 + 2),
                seq(1957 - 1, 1957 + 2),
                seq(2020 - 1, 2020 + 2))

plot_data <- results_month %>% 
  bind_rows(
    read_rds("data/deaths_monthly.Rds") %>% 
      filter(Country == "Sweden" & Year == 2021 & Month < 7) %>% 
      select(Country, Year, Month, Date, Deaths)
  ) %>% 
  filter(Year %in% plot_years) %>% 
  mutate(Difference = ifelse(excess_month > 0, "More", "Less"),
         Percent = rel_excess_month,
         Country = fct_relevel(factor(Country), 
                               "Switzerland", "Sweden", "Spain")) %>% 
  group_by(Country, Pandemic) %>% 
  mutate(time = row_number()) %>% 
  ungroup() %>% 
  filter(!is.na(Pandemic))
```

```{r echo=FALSE}
gg <- plot_data %>% 
  mutate(xmin = as.Date(paste0(Pandemic, "-01-01")),
         xmax = as.Date(paste0(Pandemic, "-12-31"))) %>% 
  ggplot(aes(x = time)) +
  geom_hline(yintercept = 0, colour = "grey60") +
  geom_rect(aes(xmin = 13, xmax = 24,
                ymin = -Inf, ymax = Inf),
            fill = "grey90", alpha = 0.7, inherit.aes = FALSE) +
  geom_col(aes(y = excess_month, fill = Difference), size = 0.5) +
  facet_grid(vars(Country), vars(Pandemic), scales = "free") + 
  scale_x_continuous(limits = c(min(plot_data$time)-1, max(plot_data$time)+1),
                     breaks = c(1, 13, 24, 36, 48), 
                     labels = c("-12m", "Start", "End", "+12m", "+24m")) +
  xlab("") + 
  ylab("Excess Deaths Per Month, n") +
  ggtitle("Excess Deaths Per Month") +
  scale_y_continuous() +
  scale_fill_manual(values = c("#67A9CF", "#EF8A62")) +  
  guides(fill = guide_legend(reverse = TRUE)) +
  theme_bw() +
  theme(panel.grid.major.x = element_blank(), 
        panel.grid.minor.x = element_blank(),
        panel.grid.major.y = element_blank(), 
        panel.grid.minor.y = element_blank(),
        strip.background = element_blank(),
        strip.text.x = element_text(margin = margin(0, 0, 2, 0)),
        strip.text.y = element_text(margin = margin(0, 0, 2, 0)),
        panel.background = element_blank(), 
        panel.border = theme_border(type = c("bottom", "left")),
        axis.line = element_line(colour = "grey40"),
        plot.margin = margin(0.5, 0.5, 0, 0))

g_data_grob = ggplotGrob(gg)

# View(g_data_grob$layout)
# gtable::gtable_show_layout(ggplotGrob(gg)) 

## remove unused facets
idx <- which(g_data_grob$layout$name %in% c("panel-3-1"))
for (i in idx) g_data_grob$grobs[[i]] <- nullGrob()

## move y axis right
# axis-l-1 needs to move 1 column to the right
idx <- which(g_data_grob$layout$name %in% c("axis-l-3"))
g_data_grob$layout[idx, c("l", "r")] <- g_data_grob$layout[idx, c("l", "r")] + 2

## move x axis up
# axis-b-1 needs to move 1 column up
idx <- which(g_data_grob$layout$name %in% c("axis-b-1"))
g_data_grob$layout[idx, c("t")] <- g_data_grob$layout[idx, c("t")] - 2

grid.newpage()
grid.draw(g_data_grob)

gg = ggplotify::as.ggplot(g_data_grob)
```

```{r include=FALSE}
ggsave(plot = gg,
       "paper/Figure_S3a.png", 
       dpi = 300, width = 297, height = 210, units = "mm")
```

### Percentage

```{r echo=FALSE}
gg <- plot_data %>% 
  mutate(xmin = as.Date(paste0(Pandemic, "-01-01")),
         xmax = as.Date(paste0(Pandemic, "-12-31"))) %>% 
  ggplot(aes(x = time)) +
  geom_hline(yintercept = 0, colour = "grey60") +
  geom_rect(aes(xmin = 13, xmax = 24,
                ymin = -Inf, ymax = Inf),
            fill = "grey90", alpha = 0.7, inherit.aes = FALSE) +
  geom_col(aes(y = Percent, fill = Difference), size = 0.5) +
  facet_grid(vars(Country), vars(Pandemic), scales = "free") + 
  scale_x_continuous(limits = c(min(plot_data$time)-1, max(plot_data$time)+1),
                     breaks = c(1, 13, 24, 36, 48), 
                     labels = c("-12m", "Start", "End", "+12m", "+24m")) +
  xlab("") + 
  ylab("Excess Deaths Per Month, %") +
  ggtitle("% Excess Deaths Per Month") +
  scale_y_continuous(labels = percent) +
  scale_fill_manual(values = c("#67A9CF", "#EF8A62")) +  
  guides(fill = guide_legend(reverse = TRUE)) +
  theme_bw() +
  theme(panel.grid.major.x = element_blank(), 
        panel.grid.minor.x = element_blank(),
        panel.grid.major.y = element_blank(), 
        panel.grid.minor.y = element_blank(),
        strip.background = element_blank(),
        strip.text.x = element_text(margin = margin(0, 0, 2, 0)),
        strip.text.y = element_text(margin = margin(0, 0, 2, 0)),
        panel.background = element_blank(), 
        panel.border = theme_border(type = c("bottom", "left")),
        axis.line = element_line(colour = "grey40"),
        plot.margin = margin(0.5, 0.5, 0, 0))

g_data_grob = ggplotGrob(gg)

# View(g_data_grob$layout)
# gtable::gtable_show_layout(ggplotGrob(gg)) 

## remove unused facets
idx <- which(g_data_grob$layout$name %in% c("panel-3-1"))
for (i in idx) g_data_grob$grobs[[i]] <- nullGrob()

## move y axis right
# axis-l-1 needs to move 1 column to the right
idx <- which(g_data_grob$layout$name %in% c("axis-l-3"))
g_data_grob$layout[idx, c("l", "r")] <- g_data_grob$layout[idx, c("l", "r")] + 2

## move x axis up
# axis-b-1 needs to move 1 column up
idx <- which(g_data_grob$layout$name %in% c("axis-b-1"))
g_data_grob$layout[idx, c("t")] <- g_data_grob$layout[idx, c("t")] - 2

grid.newpage()
grid.draw(g_data_grob)

gg = ggplotify::as.ggplot(g_data_grob)
```

```{r include=FALSE}
ggsave(plot = gg,
       "paper/Figure_S3b.png", 
       dpi = 300, width = 297, height = 210, units = "mm")

rm(plot_years, plot_data)
```

## Figure S4

Age differences across major pandemics - with %s

```{r echo=FALSE}
pandemic_years <- c(1890, 1918, 1957, 2020, 2021)

deaths_yearly_age_sex <- read_rds("data/deaths_yearly_age_sex.Rds")

gap = .3

results_age %>% 
  filter(Year %in% pandemic_years) %>%
  filter(!(Year<1957 & Age_cat=="[90,120]")) %>% 
  mutate(Age_cat2 = as.numeric(Age_cat)) %>%
  mutate(Age_cat2 = case_when(Country == "Spain" ~ Age_cat2-gap,
                              Country == "Sweden" ~ Age_cat2,
                              Country == "Switzerland" ~ Age_cat2+gap)) %>% 
  left_join(deaths_yearly_age_sex) %>%
  ggplot() +
  geom_line(aes(x = Age_cat2,
                y = rel_excess_grouped_deaths,
                colour = factor(Country))) +
  geom_pointrange(aes(x = Age_cat2,
                      y = rel_excess_grouped_deaths,
                      ymin = rel_excess_grouped_deaths_lower,
                      ymax = rel_excess_grouped_deaths_upper,
                      colour = factor(Country)),
                  size = .25) +
  
  geom_hline(yintercept = 0, linetype = 2, colour = "grey90") +
  facet_wrap(~Year, nrow = 2) +
  scale_x_continuous(labels = unique(as.character(results_age$Age_cat)),
                     breaks = 1:10) +
  scale_y_continuous(labels = percent) +
  theme_bw() +
  scale_colour_manual(values = c("darkgoldenrod1", "dodgerblue", "firebrick2")) +
  labs(x = "Year",
       y = "Montly deaths") +
  theme(legend.position = c(.85, .2),
        legend.background = element_blank(),
        strip.background = element_blank(),
        strip.text.x = element_text(margin = margin(0, 0, 2, 0)),
        axis.text.x = element_text(angle = 45, hjust = 1),
        panel.border = theme_border(type = c("bottom", "left")),
        panel.grid.major = element_blank(), 
        panel.grid.minor = element_blank(),
        plot.margin = margin(0, 0, 0, 0)) +
  labs(x = "Age group, y",
       y = "Absolute Excess Deaths, %",
       colour = "Country")
```

```{r include=FALSE}
ggsave("paper/Figure_S4.png", 
       dpi = 300, width = 280, height = 100, units = "mm")
```

## Figure S5 {.tabset}

Different age profiles of excess deaths across two pandemics and three countries.  

### Percentage

```{r echo=FALSE}
pandemic_years <- c(1918, 2020)

gap = .3

results_age %>% 
  filter(Year %in% pandemic_years) %>%
  mutate(Age_cat2=as.numeric(Age_cat)) %>%
  mutate(Age_cat2=case_when(Country=="Spain" ~ Age_cat2-gap,
                            Country=="Sweden" ~ Age_cat2,
                            Country=="Switzerland" ~ Age_cat2+gap)) %>% 
  left_join(deaths_yearly_age_sex) %>%
  ggplot() +
  geom_line(aes(x=Age_cat2,
                y=rel_excess_grouped_deaths,
                colour=factor(Year))) +
  geom_pointrange(aes(x=Age_cat2,
                      y=rel_excess_grouped_deaths,
                      ymin=rel_excess_grouped_deaths_lower,
                      ymax=rel_excess_grouped_deaths_upper,
                      colour=factor(Year)),
                  size=.25) +
  geom_hline(yintercept=0,linetype=2) +
  facet_wrap(~Country,nrow=1) +
  scale_x_continuous(labels=unique(as.character(results_age$Age_cat)),
                     breaks=1:10) +
  scale_y_continuous(labels= label_number(accuracy = 1, 
                                          suffix = "%", 
                                          scale = 100)) +
  theme_bw() +
  scale_colour_manual(values=c("darkgoldenrod1","dodgerblue","firebrick2")) +
  labs(x="Year",
       y="Montly deaths") +
  theme(legend.position = c(.1,.8),
        legend.background = element_blank(),
        axis.text.x = element_text(angle=45, hjust=1),
        plot.margin = margin(0, 0, 0, 0)) +
  labs(x="Age group",
       y="Relative excess deaths",
       colour="Country")
```

```{r include=FALSE}
ggsave("paper/Figure_S5a.png", 
       dpi = 300, width = 280, height = 100, units = "mm")
```

### Per population 

```{r echo=FALSE}
pandemic_years <- c(1918,2020)

results_age %>% 
  filter(Year %in% pandemic_years) %>%
  mutate(Age_cat2=as.numeric(Age_cat)) %>%
  mutate(Age_cat2=case_when(Country=="Spain" ~ Age_cat2-gap,
                            Country=="Sweden" ~ Age_cat2,
                            Country=="Switzerland" ~ Age_cat2+gap)) %>% 
  left_join(deaths_yearly_age_sex) %>%
  ggplot() +
  geom_line(aes(x=Age_cat2,
                y=excess_grouped_deaths/Population_obs*100000,
                colour=factor(Year))) +
  geom_pointrange(aes(x=Age_cat2,
                      y=excess_grouped_deaths/Population_obs*100000,
                      ymin=excess_grouped_deaths_lower/Population_obs*100000,
                      ymax=excess_grouped_deaths_upper/Population_obs*100000,
                      colour=factor(Year)),
                  size=.25) +
  
  geom_hline(yintercept=0, linetype=2) +
  facet_wrap(~Country, nrow=1) +
  scale_x_continuous(labels=unique(as.character(results_age$Age_cat)),
                     breaks=1:10) +
  scale_y_continuous(labels=function(x) number(x, big.mark=",")) +
  theme_bw() +
  scale_colour_manual(values=c("darkgoldenrod1","dodgerblue","firebrick2")) +
  # coord_cartesian(ylim=c(-600, 5000)) +
  labs(x="Year",
       y="Montly deaths") +
  theme(legend.position = c(.1, .8),
        legend.background = element_blank(),
        axis.text.x = element_text(angle=45, hjust=1),
        plot.margin = margin(0, 0, 0, 0)) +
  labs(x="Age group",
       y="Absolute excess deaths per 100,000 population",
       colour="Country")
```

```{r include=FALSE}
ggsave("paper/Figure_S5b.png", 
       dpi = 300, width = 280, height = 100, units = "mm")
```

## Figure S6

Impact of extremes: using 1918 pandemic as an example of impact on estimates from 1919 - 1923.  
Estimates with extreme year (blue) and without (red).  
Converge after 5 year smoothing window stops having effect (ie. in 1924).  

### Figure 1918-1925

```{r echo=FALSE}
extreme_data <- 
  left_join(
    
    results_month %>% 
      select(Country, Year, Month, Date, Deaths, 
             pred_total_deaths, 
             pred_total_deaths_lower, pred_total_deaths_upper),
    
    results_month_pand %>% 
      select(Country, Year, Month, 
             pred_total_deaths, 
             pred_total_deaths_lower, pred_total_deaths_upper) %>% 
      rename(pred_pand = pred_total_deaths,
             lower_pand = pred_total_deaths_lower, 
             upper_pand = pred_total_deaths_upper)
    
  ) %>% 
  filter(Year >= 1918 & Year <= 1924) 

extreme_data %>% 
  ggplot(aes(x = Date)) +
  geom_line(aes(y = Deaths), colour = "grey60") + 
  # with pandemics
  geom_ribbon(aes(ymin = lower_pand, ymax = upper_pand),
              alpha = 0.1, fill = "#67A9CF") +
  geom_line(aes(y = pred_pand), colour = "#67A9CF") +
  # without
  geom_ribbon(aes(ymin = pred_total_deaths_lower, 
                  ymax = pred_total_deaths_upper), 
              alpha = 0.1, fill = "#EF8A62") + 
  geom_line(aes(y = pred_total_deaths), colour = "#EF8A62") +
  scale_x_date(date_breaks = "1 year", date_labels = "%Y") +
  # coord_cartesian(ylim = c(NA, 100000)) + 
  xlab("") + ylab("Number of Deaths") + 
  facet_grid(vars(Country), scales = "free_y") + 
  theme_bw() + 
  theme(plot.margin = margin(0.5, 0.5, 0, 0))
```

```{r include=FALSE}
ggsave("paper/Figure_S6a.png", 
       dpi = 300, width = 297, height = 210, units = "mm")
```

### Table 1919

Using 1918 pandemic as an example of impact on estimates from 1919.   

```{r echo=FALSE}
extreme_data %<>% 
  filter(Year == 1919) %>% 
  select(-pred_total_deaths_lower, -pred_total_deaths_upper, 
         -lower_pand, -upper_pand, 
         -Date, -Year) %>% 
  mutate(diff_1 = Deaths - pred_total_deaths,
         diff_2 = Deaths - pred_pand) %>% 
  relocate(diff_1, .after = pred_total_deaths) 

extreme_data %>% 
  kbl(col.names = c("Country", "Month", "Observed", 
                    "Predicted", "Difference",
                    "Predicted (pandemic)", "Difference")) %>%
  kable_styling(bootstrap_options = c("striped", "hover", "condensed"),
                full_width = FALSE) %>% 
  column_spec(1, italic = TRUE) %>%
  column_spec(5, color = ifelse(extreme_data$diff_1 > 0, 
                                "#EF8A62", "#67A9CF")) %>% 
  column_spec(7, color = ifelse(extreme_data$diff_2 > 0, 
                                "#EF8A62", "#67A9CF")) %>% 
  row_spec(1, italic = FALSE) %>% 
  add_header_above(c(" " = 3, "Including 1918" = 2, "Excluding 1918" = 2)) %>%
  # collapse not working; issue filled on github
  collapse_rows(columns = 1, valign = "top")
```

### Figure 2020-21

```{r echo=FALSE}
extreme_data <- 
  left_join(
    
    results_month %>% 
      select(Country, Year, Month, Date, Deaths, 
             pred_total_deaths, 
             pred_total_deaths_lower, pred_total_deaths_upper),
    
    results_month_pand %>% 
      select(Country, Year, Month, 
             pred_total_deaths, 
             pred_total_deaths_lower, pred_total_deaths_upper) %>% 
      rename(pred_pand = pred_total_deaths,
             lower_pand = pred_total_deaths_lower, 
             upper_pand = pred_total_deaths_upper)
    
  ) %>% 
  filter(Year >= 2020) %>% 
  filter(Country != "Sweden")

extreme_data %>% 
  ggplot(aes(x = Date)) +
  geom_line(aes(y = Deaths), colour = "grey60") + 
  # with pandemics
  geom_ribbon(aes(ymin = lower_pand, ymax = upper_pand),
              alpha = 0.1, fill = "#67A9CF") +
  geom_line(aes(y = pred_pand), colour = "#67A9CF") +
  # without
  geom_ribbon(aes(ymin = pred_total_deaths_lower, 
                  ymax = pred_total_deaths_upper), 
              alpha = 0.1, fill = "#EF8A62") + 
  geom_line(aes(y = pred_total_deaths), colour = "#EF8A62") +
  scale_x_date(date_breaks = "1 year", date_labels = "%Y") +
  # coord_cartesian(ylim = c(NA, 100000)) + 
  xlab("") + ylab("Number of Deaths") + 
  facet_grid(vars(Country), scales = "free_y") + 
  theme_bw() + 
  theme(plot.margin = margin(0.5, 0.5, 0, 0))
```

```{r include=FALSE}
ggsave("paper/Figure_S6b.png", 
       dpi = 300, width = 297, height = 210, units = "mm")
```

### Table 2021

Using 2021 pandemic as an example of impact on estimates from 2020   

```{r echo=FALSE}
extreme_data %<>% 
  filter(Year == 2021) %>% 
  select(-pred_total_deaths_lower, -pred_total_deaths_upper, 
         -lower_pand, -upper_pand, 
         -Date, -Year) %>% 
  mutate(diff_1 = Deaths - pred_total_deaths,
         diff_2 = Deaths - pred_pand) %>% 
  relocate(diff_1, .after = pred_total_deaths) %>% 
  filter(Country != "Sweden")

extreme_data %>% 
  kbl(col.names = c("Country", "Month", "Observed", 
                    "Predicted", "Difference",
                    "Predicted (pandemic)", "Difference")) %>%
  kable_styling(bootstrap_options = c("striped", "hover", "condensed"),
                full_width = FALSE) %>% 
  column_spec(1, italic = TRUE) %>%
  column_spec(5, color = ifelse(extreme_data$diff_1 > 0, 
                                "#EF8A62", "#67A9CF")) %>% 
  column_spec(7, color = ifelse(extreme_data$diff_2 > 0, 
                                "#EF8A62", "#67A9CF")) %>% 
  row_spec(1, italic = FALSE) %>% 
  add_header_above(c(" " = 3, "Including 1918" = 2, "Excluding 1918" = 2)) %>%
  # collapse not working; issue filled on github
  collapse_rows(columns = 1, valign = "top")
```

## Pandemic compared to last year {.tabset}

**2021 not included yet since not a full year!**  

```{r echo=FALSE}
pandemic_years <- c(1890, 1918, 1957, 2020)

analysis_years <- c(seq(1890 - 10, 1890),
                    seq(1918 - 10, 1918),
                    seq(1957 - 10, 1957),
                    seq(2020 - 10, 2020))

previous_years <- c(seq(1890 - 10, 1890 - 1),
                    seq(1918 - 10, 1918 - 1),
                    seq(1957 - 10, 1957 - 1),
                    seq(2020 - 10, 2020 - 1))

age_plot_data <- 
  results_age %>% 
  filter(Year %in% analysis_years) %>% 
  mutate(Type = ifelse(Year %in% previous_years, "Previous", "")) %>% 
  mutate(Type = ifelse(Year %in% pandemic_years, "Pandemic", Type)) %>%  
  arrange(Country, Year, Age_cat) %>% 
  mutate(Pandemic = ifelse(Year %in% pandemic_years, Year, NA)) %>% 
  fill(Pandemic, .direction = "up") 

previous <- pandemic_years - 1

data_year <- left_join(
  
  age_plot_data %>% 
    filter(Type == "Pandemic") %>% 
    select(Pandemic, Year, Country, Age_cat, Deaths) %>% 
    rename(Deaths_pandemic = Deaths),
  
  age_plot_data %>% 
    filter(Year %in% previous) %>% 
    select(Pandemic, Year, Country, Age_cat, Deaths) %>% 
    rename(Deaths_previous = Deaths) %>% 
    mutate(Year = Year + 1)
  
) %>% 
  mutate(Exc = Deaths_pandemic - Deaths_previous,
         Exc_perc = (Deaths_pandemic - Deaths_previous) / Deaths_previous)

```

### Absolute

```{r echo=FALSE}
ggplot(data = data_year,
       aes(x = Age_cat, y = Exc)) + 
  geom_col() +
  facet_grid(rows = vars(Country), cols = vars(Pandemic), scales = "free_y") + 
  scale_y_continuous(labels = label_number(accuracy = 0.1, suffix = "K", scale = 1e-4)) +
  theme_minimal() + 
  theme(axis.text.x = element_text(angle = 90)) +
  xlab("Age group") + 
  ylab("Absolute difference in deaths to previous year")
```

### Relative

```{r echo=FALSE}
ggplot(data = data_year,
       aes(x = Age_cat, y = Exc_perc)) + 
  geom_col() +
  facet_grid(rows = vars(Country), cols = vars(Pandemic)) + 
  scale_y_continuous(labels = percent) +
  theme_minimal() + 
  theme(axis.text.x = element_text(angle = 90)) +
  xlab("Age group") + 
  ylab("% difference in deaths to previous year")
```


## Comparing to BfS estimates

Data from BfS prepared in `03`. No CI included in comparisons since different age groups are used!  

```{r echo=FALSE}
excess_2010_2020 <- 
  left_join(
    # official counts
    read_rds("data/BfS/excess_2010_2020.Rds") %>% 
      rename(Excess_BfS = Excess_expected),
    
    # our estimates
    results_year %>% 
      filter(Country == "Switzerland") %>% 
      select(Year, yearly_excess_total_deaths) %>% 
      rename(Excess_modelled = yearly_excess_total_deaths)
  ) %>% 
  mutate(Difference = Excess_modelled - Excess_BfS)
```

```{r echo=FALSE}
excess_2010_2020 %>% 
  kbl() %>%
  kable_styling(bootstrap_options = c("striped", "hover", "condensed"),
                full_width = FALSE) %>% 
  column_spec(1, italic = TRUE) %>%
  row_spec(1, italic = FALSE) %>% 
  column_spec(4, color = ifelse(excess_2010_2020$Difference > 0, 
                                "#EF8A62", "#67A9CF"))
```


<!-- ----------------------------------------------------- -->

# Model diagnostics

## Rhat

Check that all values <1.1

```{r}
summary(results_month[results_month$Model != "Global Serfling", ]$Rhat) 
```

## ESS

Check that all effective sample sizes are large

```{r}
summary(results_month[results_month$Model != "Global Serfling", ]$n_eff) 
```

## Model agreement

Example of 2020    

```{r echo=FALSE}
bind_rows(
  read_rds("data/outputs_2021-11-18/All_results_month.Rds"), 
  read_rds("data/outputs_2021-11-23/Switzerland_results_month.Rds"), 
  read_rds("data/outputs_2021-11-23/Sweden_results_month.Rds"), 
  read_rds("data/outputs_2021-11-24/Spain_results_month.Rds")
) %>%  
  filter(Year == 2020) %>% 
  ggplot() +
  geom_pointrange(aes(x=Month, y=excess_month, 
                      ymin=excess_month_lower, ymax=excess_month_upper,
                      colour=Model),
                  position=position_dodge(.8)) +
  facet_grid(Country~Year, scales="free") + 
  theme_bw()
```

And 1918  

```{r echo=FALSE}
bind_rows(
  read_rds("data/outputs_2021-11-18/All_results_month.Rds"), 
  read_rds("data/outputs_2021-11-23/Switzerland_results_month.Rds"), 
  read_rds("data/outputs_2021-11-23/Sweden_results_month.Rds"), 
  read_rds("data/outputs_2021-11-24/Spain_results_month.Rds")
) %>%  
  filter(Year == 1918) %>% 
  ggplot() +
  geom_pointrange(aes(x=Month, y=excess_month, 
                      ymin=excess_month_lower, ymax=excess_month_upper,
                      colour=Model),
                  position=position_dodge(.8)) +
  facet_grid(Country~Year, scales="free") + 
  theme_bw()
```

## Check age

```{r echo=FALSE}
results_age %>% 
  filter(Year >= 2010) %>% 
  ggplot() +
  geom_pointrange(aes(x=Year, 
                      y=yearly_excess_grouped_deaths,
                      ymin=yearly_excess_grouped_deaths_lower,
                      ymax=yearly_excess_grouped_deaths_upper),
                  position=position_dodge(.8)) +
  facet_grid(Country~Age_cat) + 
  theme_bw()
```

```{r echo=FALSE}
results_age %>% 
  filter(Year >= 1914 & Year <= 1922) %>% 
  ggplot() +
  geom_pointrange(aes(x=Year, 
                      y=yearly_excess_grouped_deaths,
                      ymin=yearly_excess_grouped_deaths_lower,
                      ymax=yearly_excess_grouped_deaths_upper),
                  position=position_dodge(.8)) +
  facet_grid(Country~Age_cat) + 
  theme_bw()
```

<!-- ----------------------------------------------------- -->

```{r echo=FALSE, eval=FALSE}
# Alt Fig 3
ggplot(data = age_plot_data %>% 
         filter(Type == "Previous"),
       aes(x = Age_cat)) + 
  
  # observed from previous decade
  geom_boxplot(aes(y = Deaths,
                   fill = "grey"), 
               color = "grey", alpha = 0.25) +
  geom_jitter(aes(y = Deaths), 
              color = "black", size = 0.5, alpha = 0.5) +
  
  # modelled 67A9CF
  geom_ribbon(data = age_plot_data %>% 
                filter(Type == "Pandemic"),
              aes(ymin = pred_grouped_deaths_lower, 
                  ymax = pred_grouped_deaths_upper, 
                  group = Year), 
              alpha = 0.1, fill = "#67A9CF") + 
  geom_line(data = age_plot_data %>% 
              filter(Type == "Pandemic"),
            aes(y = pred_grouped_deaths, 
                group = Year, colour = "#67A9CF")) +
  
  # observed EF8A62
  geom_line(data = age_plot_data %>% 
              filter(Type == "Pandemic"), 
            aes(y = Deaths, group = Year,
                colour = "#EF8A62")) +
  
  facet_grid(rows = vars(Country), cols = vars(Pandemic), scales = "free_y") + 
  scale_y_continuous(labels = label_number(accuracy = 0.1, suffix = "K", scale = 1e-4)) +
  scale_fill_manual(name = NULL,
                    values = c("grey" = "grey"), 
                    labels = c("Observed \n(Last \n10 years)"), 
                    guide = "legend") +  
  scale_colour_manual(name = "Deaths",
                      values = c("#67A9CF" = "#67A9CF",
                                 "#EF8A62" = "#EF8A62"), 
                      labels = c("Expected \n(Last \n5 years)",
                                 "Observed"), 
                      guide = "legend") +
  theme_bw() +
  theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
        panel.background = element_blank(), 
        axis.line = element_line(colour = "grey40"),
        axis.ticks = element_line(),
        axis.text.x = element_text(angle = 90)) +
  xlab("Age group") + labs(colour = "Age group") +
  guides(fill = guide_legend(order = 2),
         colour = guide_legend(order = 1))

# ggsave("paper/Figure_3.png", dpi = 300, width = 297, height = 210, units = "mm")
```

<!-- ----------------------------------------------------- -->

# Computing Environment

`r mu$session()`
