==> analyses/06_comparison_week_month.Rmd <==
---
title: "Excess mortality"
subtitle: "2005-2020 Comparison of monthly and weekly estimates"
author: "Katarina Matthes, Radek Panczak"
date: "`r Sys.Date()`"
output:
  rmdformats::robobook:
    code_folding: show
    self_contained: yes
    highlight: pygments
editor_options: 
  chunk_output_type: console
knit: (function(inputFile, encoding) {
  rmarkdown::render(inputFile, encoding = encoding, output_dir = "../docs") })
---

<style type="text/css">
div.page-inner {
  max-width: 10000px !important;
}
</style>

<!-- ----------------------------------------------------- -->

```{r setup, include = FALSE}
set.seed(12345)
options(scipen = 999)
options(max.print = "75")
```

```{r knit-setup, include = FALSE}
knitr::opts_knit$set(root.dir = rprojroot::find_rstudio_root_file())

knitr::opts_chunk$set(cache = FALSE,
                      prompt = FALSE,
                      tidy = FALSE,
                      comment = NA,
                      message = FALSE,
                      warning = FALSE)

knitr::opts_knit$set(width = 75)

mu <- Hmisc::markupSpecs$html
```

<!-- ----------------------------------------------------- -->

```{r global_options, include=FALSE}
source("R/fn_comparison_monthly_weekly.R")
CH <- fn_comparison_monthly_weekly("Switzerland")
SW <- fn_comparison_monthly_weekly("Sweden")
ES <- fn_comparison_monthly_weekly("Spain")
```

# Explanation

There are differences between months and weeks. Months go exactly to 31.December, whereas calendar weeks go into the new year. 2015 and 2020 had even 53 calendar weeks, instead of 52. Which means, the KW 1 of the 2020 started on `30.12.2019` and KW 53 ended on `03.01.2021`, i.e. 5 days did not actually belong to the year 2020. Calendar weeks are based on a DIN standard. But since we want to represent years, I would take the calculation with months.There we have then in any case always the 01.01. - 31.12. in it and no fluctuations of the days by the DIN norm of the calendar weeks.

Because of these differences, even the reported death rates of months and weeks vary. As said, for calendar weeks, 5 more days are counted. And if the reported death numbers are already different, then the expected numbers are as well. I can't make any statement right now about the method, whether the calculation differs, because we already assume a different basis at the beginning. 

# Switzerland {.tabset}

```{r table_Switzerland, echo=FALSE,message=TRUE} 
CH %>%
  kbl() %>%
  kable_material(c("striped", "hover"),full_width = F,position = "left")
```

## Absolute difference 

```{r echo=FALSE}
CH %>% 
  mutate(Year=as.integer(as.character(Year))) %>%
  ggplot(aes(x = Year, y = Difference)) +
  geom_line()
```

## % difference 

```{r echo=FALSE}
CH %>% 
  mutate(Year=as.integer(as.character(Year))) %>%
  ggplot(aes(x = Year, y = Perc_diff)) +
  geom_line()
```

# Sweden {.tabset}

```{r table_Sweden, echo=FALSE,message=TRUE} 
SW %>%
  kbl() %>%
  kable_material(c("striped", "hover"),full_width = F,position = "left")
```

## Absolute difference 

```{r echo=FALSE}
SW %>% 
  mutate(Year=as.integer(as.character(Year))) %>%
  ggplot(aes(x = Year, y = Difference)) +
  geom_line()
```

## % difference 

```{r echo=FALSE}
SW %>% 
  mutate(Year=as.integer(as.character(Year))) %>%
  ggplot(aes(x = Year, y = Perc_diff)) +
  geom_line()
```

# Spain {.tabset}

```{r table_Spain, echo=FALSE,message=TRUE} 
ES %>%
  kbl() %>%
  kable_material(c("striped", "hover"),full_width = F,position = "left")
```

## Absolute difference 

```{r echo=FALSE}
ES %>% 
  mutate(Year=as.integer(as.character(Year))) %>%
  ggplot(aes(x = Year, y = Difference)) +
  geom_line()
```

## % difference 

```{r echo=FALSE}
ES %>% 
  mutate(Year=as.integer(as.character(Year))) %>%
  ggplot(aes(x = Year, y = Perc_diff)) +
  geom_line()
```
